feat(grow→polish): migrate narrative_gaps to POLISH Phase 1a (#1368, PR A)
First of three migration PRs (per issue #1368) implementing the code
half of the GROW→POLISH spec migration that landed in PR #1366.

GROW Phase 4b (narrative_gaps) is narrative-prep work, not structural
skeleton, so per the structural-vs-narrative principle it belongs in
POLISH. This PR moves the implementation; the spec already documents
the new home as POLISH Phase 1a.

## What changed

**New shared graph module** (`src/questfoundry/graph/gap_insertion.py`):
- `GapInsertionReport` dataclass (moved from grow/llm_helper.py)
- `validate_and_insert_gaps()` free function — full gap-validation +
  insertion logic, including cycle prevention, path-membership
  checks, ID normalization, and ordering checks. Single source of
  truth for gap insertion now that two stages need it.

**GROW** (`grow/llm_helper.py`, `grow/llm_phases.py`):
- `_validate_and_insert_gaps` is now a thin delegation to the free
  function (back-compat for existing test call sites).
- `_phase_4b_narrative_gaps` removed; pacing_gaps' `depends_on` now
  references `scene_types` directly (was `narrative_gaps`).
- GROW Phase 4 in code dropped from 7 to 6 sub-phases.

**POLISH** (`polish/llm_phases.py`):
- New `_phase_1a_narrative_gaps` decorated with
  `@polish_phase(name="narrative_gaps", depends_on=["beat_reordering"], priority=1)`.
- Calls `validate_and_insert_gaps` from the shared module.
- `pacing` phase's `depends_on` updated from `["beat_reordering"]`
  to `["narrative_gaps"]` (mirrors the polish.md ordering).
- Logs structured `phase1a_complete` event.

**Prompt** (`prompts/templates/polish_phase1a_narrative_gaps.yaml`):
- New template, identical body to `grow_phase4b_narrative_gaps.yaml`
  except for the `name`/`description` header. Old GROW template
  remains in place since pacing_gaps (still in GROW) doesn't share
  it; will be cleaned up in PR C when pacing_gaps moves.

**Tests**:
- `TestPhase4bNarrativeGaps` removed from `test_grow_stage.py`
  (with a NOTE pointing to the new location).
- Equivalent `TestPolishPhase1aNarrativeGaps` added to
  `test_polish_llm_phases.py` covering: insert valid gap, skip
  invalid path, no-paths early-return (now `skipped` per POLISH
  convention), single-beat-path early-return.
- `test_grow_stage.py::TestGrowStagePhaseOrder`: phase count 18 → 17.
- `test_grow_registry.py::TestGlobalRegistry`: same — 18 → 17 phases.

## Future PRs in the same epic

- **PR B**: atmospheric → POLISH Phase 5e + path_arcs → POLISH Phase 5f.
- **PR C**: pacing_gaps → POLISH Phase 2 extension + entity_arcs →
  POLISH Phase 3 extension. Will also remove the temporary
  `_validate_and_insert_gaps` delegation on `_LLMHelperMixin` once
  GROW no longer needs it.

Tests: 3566 unit tests pass; mypy + ruff clean.

Refs #1368.
