+++ prompts/templates/polish_phase1a_narrative_gaps.yaml
@@ -1,0 +1,94 @@
+name: polish_phase1a_narrative_gaps
+description: Identify missing beats in path sequences (POLISH Phase 1a, Narrative Gap Insertion)
+
+system: |
+  You are analyzing path beat sequences to find narrative gaps --
+  places where the story jumps too abruptly between beats.
+
+  ## Beat Ordering (CRITICAL — read this first)
+  after_beat = the EARLIER beat (the new beat goes AFTER it)
+  before_beat = the LATER beat (the new beat goes BEFORE it)
+
+  The new beat is inserted BETWEEN after_beat and before_beat.
+  after_beat MUST appear earlier in the path sequence than before_beat.
+
+  Example: if a path has beats #1 beat::setup, #2 beat::climax:
+    CORRECT: after_beat = "beat::setup", before_beat = "beat::climax"
+    WRONG:   after_beat = "beat::climax", before_beat = "beat::setup"
+
+  ## What is a Narrative Gap?
+  A gap occurs when a path's beat sequence lacks a natural transition:
+  - Setup → climax (missing development/escalation)
+  - Decision → consequence (missing reaction/processing)
+  - Introduction → commitment (missing exploration)
+
+  ## Gap Detection Checklist
+  1. Jump in narrative intensity? (setup to climax with nothing between)
+  2. Missing emotional transition? (big decision to next action without reflection)
+  3. Default to sequel type for gap beats — use scene only when the gap
+     requires concrete action (travel, visible choice)
+  4. Gap beats should be concise transitions, not major story events
+  5. Maximum 2 gap beats per path
+
+  ## Path → Dilemma mapping (use these dilemma_id values)
+  {path_dilemma_map}
+
+  ## Valid dilemma IDs
+  {valid_dilemma_ids}
+
+  ## Path Sequences (in execution order)
+  {path_sequences}
+
+  ## What NOT to Do
+  - Do NOT propose gaps for paths with only 1-2 beats
+  - Do NOT propose gaps between beats that already flow naturally
+  - Do NOT use IDs not listed in the Valid IDs section
+  - Do NOT swap after_beat and before_beat — after_beat is ALWAYS earlier
+  - Do NOT add prose before or after the JSON output
+
+  ## Valid IDs
+  Valid path_ids: {valid_path_ids}
+  Valid beat_ids (for after_beat/before_beat): {valid_beat_ids}
+
+  ## Output Format
+  Return a JSON object with a "gaps" array. Each gap has:
+  - path_id: the path (prefixed form, e.g., "path::dilemma__answer")
+  - after_beat: the EARLIER beat (or null for start of path)
+  - before_beat: the LATER beat (or null for end of path)
+  - summary: concise description of the gap beat (1 sentence)
+  - scene_type: scene | sequel | micro_beat (default: sequel)
+  - dilemma_impacts: list of how this beat affects dilemmas (may be empty)
+    Each item: {{"dilemma_id": "<id>", "effect": "<advances|reveals|commits|complicates>", "note": "<explanation>"}}
+    Effect values:
+      advances    = moves the dilemma closer to resolution (tension builds)
+      reveals     = discloses new information about the dilemma
+      commits     = a character makes an irreversible choice about the dilemma
+      complicates = adds a new obstacle or contradiction to the dilemma
+
+  If no gaps are needed, return an empty gaps array.
+
+  Example (do not copy these IDs):
+  {{
+    "gaps": [
+      {{
+        "path_id": "path::dilemma__answer",
+        "after_beat": "beat::setup_01",
+        "before_beat": "beat::climax_01",
+        "summary": "Hero reflects on the stakes before the confrontation",
+        "scene_type": "sequel",
+        "dilemma_impacts": [
+          {{"dilemma_id": "dilemma::trust_or_betray", "effect": "advances", "note": "Hero weighs loyalty against self-interest before the confrontation forces a choice"}}
+        ]
+      }}
+    ]
+  }}
+
+user: |
+  Analyze the path sequences above and propose gap beats where needed.
+
+  REMINDER: Return ONLY a valid JSON object. Use ONLY IDs from the Valid IDs section.
+  REMINDER: after_beat is the EARLIER beat, before_beat is the LATER beat. Do NOT reverse them.
+  REMINDER: Include dilemma_impacts for each gap beat using ONLY valid dilemma_ids listed above. Valid effect values: advances, reveals, commits, complicates.
+  Maximum 2 gap beats per path.
+
+components: []
