+++ tests/unit/test_polish_llm_phases.py
@@ -2,11 +2,16 @@
 
 from __future__ import annotations
 
+from unittest.mock import AsyncMock, MagicMock
+
+import pytest
+
 from questfoundry.graph.graph import Graph
 from questfoundry.pipeline.stages.polish.llm_phases import (
     _detect_duplicate_labels_in_passage,
     _detect_pacing_flags,
 )
+from questfoundry.pipeline.stages.polish.stage import PolishStage
 
 
 def _make_beat(graph: Graph, beat_id: str, summary: str, **kwargs: object) -> None:
@@ -164,3 +169,113 @@ def test_detect_duplicate_labels_multiple_collisions_sorted() -> None:
     ]
     collisions = _detect_duplicate_labels_in_passage(specs)
     assert [c["from_passage"] for c in collisions] == ["passage::p1", "passage::p2"]
+
+
+# ---------------------------------------------------------------------------
+# Phase 1a Narrative Gap Insertion (migrated from GROW Phase 4b per PR #1366)
+# ---------------------------------------------------------------------------
+
+
+class TestPolishPhase1aNarrativeGaps:
+    """POLISH Phase 1a — Narrative Gap Insertion.
+
+    Migrated from ``tests/unit/test_grow_stage.py::TestPhase4bNarrativeGaps``
+    when the implementation moved per the structural-vs-narrative migration
+    (issue #1368). Behavior is identical; only the host stage changed.
+    """
+
+    @pytest.mark.asyncio
+    async def test_phase_1a_inserts_gap_beats(self) -> None:
+        """Phase 1a inserts gap beats from LLM proposals."""
+        from questfoundry.models.grow import GapProposal, Phase4bOutput
+        from tests.fixtures.grow_fixtures import make_single_dilemma_graph
+
+        graph = make_single_dilemma_graph()
+        stage = PolishStage()
+
+        phase_output = Phase4bOutput(
+            gaps=[
+                GapProposal(
+                    path_id="path::mentor_trust_canonical",
+                    after_beat="beat::mentor_meet",
+                    before_beat="beat::mentor_commits_canonical",
+                    summary="Hero reflects on mentor's words",
+                    scene_type="sequel",
+                ),
+            ]
+        )
+
+        mock_structured = AsyncMock()
+        mock_structured.ainvoke = AsyncMock(return_value=phase_output)
+        mock_model = MagicMock()
+        mock_model.with_structured_output = MagicMock(return_value=mock_structured)
+
+        result = await stage._phase_1a_narrative_gaps(graph, mock_model)
+
+        assert result.status == "completed"
+        assert result.llm_calls == 1
+        assert "1" in result.detail
+
+        beat_nodes = graph.get_nodes_by_type("beat")
+        gap_beats = [bid for bid in beat_nodes if "gap" in bid]
+        assert len(gap_beats) == 1
+
+    @pytest.mark.asyncio
+    async def test_phase_1a_skips_invalid_path(self) -> None:
+        """Phase 1a skips gap proposals with invalid path IDs."""
+        from questfoundry.models.grow import GapProposal, Phase4bOutput
+        from tests.fixtures.grow_fixtures import make_single_dilemma_graph
+
+        graph = make_single_dilemma_graph()
+        stage = PolishStage()
+
+        phase_output = Phase4bOutput(
+            gaps=[
+                GapProposal(
+                    path_id="path::nonexistent",
+                    after_beat="beat::opening",
+                    before_beat="beat::mentor_meet",
+                    summary="Invalid path gap",
+                    scene_type="sequel",
+                ),
+            ]
+        )
+
+        mock_structured = AsyncMock()
+        mock_structured.ainvoke = AsyncMock(return_value=phase_output)
+        mock_model = MagicMock()
+        mock_model.with_structured_output = MagicMock(return_value=mock_structured)
+
+        result = await stage._phase_1a_narrative_gaps(graph, mock_model)
+
+        assert result.status == "completed"
+        assert "0" in result.detail
+
+    @pytest.mark.asyncio
+    async def test_phase_1a_no_paths(self) -> None:
+        """Phase 1a returns skipped when no paths exist."""
+        graph = Graph.empty()
+        stage = PolishStage()
+        mock_model = MagicMock()
+
+        result = await stage._phase_1a_narrative_gaps(graph, mock_model)
+
+        assert result.status == "skipped"
+        assert "No paths" in result.detail
+        assert result.llm_calls == 0
+
+    @pytest.mark.asyncio
+    async def test_phase_1a_single_beat_paths_skipped(self) -> None:
+        """Phase 1a skips paths with only 1 beat (no sequence to gap-check)."""
+        graph = Graph.empty()
+        graph.create_node("path::short", {"type": "path", "raw_id": "short"})
+        graph.create_node("beat::only", {"type": "beat", "summary": "Lone beat"})
+        graph.add_edge("belongs_to", "beat::only", "path::short")
+
+        stage = PolishStage()
+        mock_model = MagicMock()
+
+        result = await stage._phase_1a_narrative_gaps(graph, mock_model)
+
+        assert result.status == "skipped"
+        assert "No paths with 2+ beats" in result.detail
